Struct bn::arith::U512

#[repr(C)]

pub struct U512(pub [u64; 8]);

512-bit, stack allocated biginteger for use in extension field serialization and scalar interpretation.

Methods

impl U512

pub fn from(c1: &U256, c0: &U256, modulo: &U256) -> U512

Multiplies c1 by modulo, adds c0.

pub fn random<R: Rng>(rng: &mut R) -> U512

Get a random U512

pub fn get_bit(&self, n: usize) -> Option<bool>

pub fn divrem(&self, modulo: &U256) -> (Option<U256>, U256)

Divides self by modulo, returning remainder and, if possible, a quotient smaller than the modulus.

pub fn interpret(buf: &[u8; 64]) -> U512